requirements.txt
mss
opencv-python
pillow
pyautogui
numpy